Move summary loading to summary response class
