+++ src/ert/_c_wrappers/enkf/enkf_state.py
@@ -1,14 +1,11 @@
 from __future__ import annotations
 
-import ctypes
 import logging
-from fnmatch import fnmatch
 from pathlib import Path
 from typing import TYPE_CHECKING
 
 import numpy as np
 import xarray as xr
-from ecl.summary import EclSum
 
 from ert._c_wrappers.enkf import SummaryConfig
 from ert.load_status import LoadResult, LoadStatus
@@ -56,91 +53,21 @@ def _internalize_GEN_DATA(
     return LoadResult(LoadStatus.LOAD_SUCCESSFUL, "")
 
 
-def _should_load_summary_key(data_key, user_set_keys):
-    return any(fnmatch(data_key, key) for key in user_set_keys)
-
-
-def _load_summary_data(run_path: str, job_name: str) -> EclSum:
-    try:
-        summary = EclSum(
-            f"{run_path}/{job_name}",
-            include_restart=False,
-            lazy_load=False,
-        )
-    except IOError as e:
-        raise IOError("Failed to load summary data from disk.") from e
-
-    return summary
-
-
-def _internalize_SUMMARY_DATA(
-    ens_config: EnsembleConfig, run_arg: RunArg, summary: EclSum, config: SummaryConfig
-) -> LoadResult:
-    data = []
-    keys = []
-    time_map = summary.alloc_time_vector(True)
-    axis = [t.datetime() for t in time_map]
-    if config.refcase:
-        existing_time_map = ens_config.refcase.alloc_time_vector(True)
-        missing = []
-        for step, (response_t, reference_t) in enumerate(
-            zip(time_map, existing_time_map)
-        ):
-            if response_t not in existing_time_map:
-                missing.append((response_t, reference_t, step + 1))
-        if missing:
-            logger.warning(
-                f"Realization: {run_arg.iens}, load warning: {len(missing)} "
-                "inconsistencies in time map, first: "
-                f"Time mismatch for step: {missing[0][2]}, response time: "
-                f"{missing[0][0]}, reference case: {missing[0][1]}, last: Time "
-                f"mismatch for step: {missing[-1][2]}, response time: {missing[-1][0]}"
-                f", reference case: {missing[-1][1]} from: {run_arg.runpath}/"
-                f"{config.input_file}.UNSMRY"
-            )
-
-    user_summary_keys = set(config.keys)
-    for key in summary:
-        if not _should_load_summary_key(key, user_summary_keys):
-            continue
-        keys.append(key)
-
-        np_vector = np.zeros(len(time_map))
-        summary._init_numpy_vector_interp(
-            key,
-            time_map,
-            np_vector.ctypes.data_as(ctypes.POINTER(ctypes.c_double)),
-        )
-        data.append(np_vector)
-
-    ds = xr.Dataset(
-        {"values": (["name", "time"], data)},
-        coords={"time": axis, "name": keys},
-    )
-    run_arg.ensemble_storage.save_response("summary", ds, run_arg.iens)
-
-    return LoadResult(LoadStatus.LOAD_SUCCESSFUL, "")
-
-
 def _write_summary_data_to_storage(
     ens_config: EnsembleConfig, run_arg: RunArg
 ) -> LoadResult:
-    user_summary_keys = ens_config.get_summary_keys()
-    if not user_summary_keys:
-        return LoadResult(LoadStatus.LOAD_SUCCESSFUL, "")
-    config = ens_config["summary"]
-    assert isinstance(config, SummaryConfig)
-    filename = config.input_file.replace("<IENS>", str(run_arg.iens))
-    try:
-        summary = _load_summary_data(run_arg.runpath, filename)
-    except IOError:
-        return LoadResult(
-            LoadStatus.LOAD_FAILURE,
-            "Could not find SUMMARY file or using non unified SUMMARY "
-            f"file from: {run_arg.runpath}/{filename}.UNSMRY",
-        )
-
-    return _internalize_SUMMARY_DATA(ens_config, run_arg, summary, config)
+    for config in ens_config.response_configs.values():
+        if isinstance(config, SummaryConfig):
+            # Nothing to load, should not be handled here, should never be
+            # added in the first place
+            if not config.keys:
+                return LoadResult(LoadStatus.LOAD_SUCCESSFUL, "")
+            try:
+                ds = config.read_from_file(run_arg.runpath, run_arg.iens)
+                run_arg.ensemble_storage.save_response("summary", ds, run_arg.iens)
+            except ValueError as err:
+                return LoadResult(LoadStatus.LOAD_FAILURE, str(err))
+    return LoadResult(LoadStatus.LOAD_SUCCESSFUL, "")
 
 
 def _write_gen_data_to_storage(
